+++ u8g2/csrc/u8g2_setup.c
@@ -22,6 +22,7 @@
 static void u8g2_update_dimension_common(u8g2_t *u8g2)
 {
   u8g2_uint_t h, w;
+  
   h = u8g2->tile_buf_height;
   
   /* handle the case, where the buffer is larger than the (remaining) part of the display */
@@ -41,7 +42,7 @@ static void u8g2_update_dimension_common(u8g2_t *u8g2)
   u8g2->buf_y1 += h;
 
   u8g2->width = w;
-  u8g2->height = h;  
+  u8g2->height = u8g2_GetU8x8(u8g2)->display_info->tile_height * 8;  
 }
 
 void u8g2_update_dimension_r0(u8g2_t *u8g2)
@@ -77,6 +78,20 @@ void u8g2_update_dimension_r1(u8g2_t *u8g2)
       u8g2->user_x0, u8g2->user_x1, u8g2->user_y0, u8g2->user_y1);
 }
 
+void u8g2_update_dimension_r2(u8g2_t *u8g2)
+{
+  u8g2_update_dimension_common(u8g2);
+
+  u8g2->user_x0 = u8g2->buf_x0;
+  u8g2->user_x1 = u8g2->buf_x1;
+  
+  u8g2->user_y0 = u8g2->height - u8g2->buf_y1;
+  u8g2->user_y1 = u8g2->height - u8g2->buf_y0;
+  
+  printf("x0=%d x1=%d y0=%d y1=%d\n", 
+      u8g2->user_x0, u8g2->user_x1, u8g2->user_y0, u8g2->user_y1);
+}
+
 
 /*============================================*/
 extern void u8g2_draw_hv_line_4dir(u8g2_t *u8g2, u8g2_uint_t x, u8g2_uint_t y, u8g2_uint_t len, uint8_t dir);
@@ -90,12 +105,31 @@ void u8g2_draw_l90_r0(u8g2_t *u8g2, u8g2_uint_t x, u8g2_uint_t y, u8g2_uint_t le
 void u8g2_draw_l90_r1(u8g2_t *u8g2, u8g2_uint_t x, u8g2_uint_t y, u8g2_uint_t len, uint8_t dir)
 {
   u8g2_uint_t xx, yy;
+  
   yy = x;
-  xx = u8g2_GetU8x8(u8g2)->display_info->tile_width;
-  xx *= 8;
+  
+  xx = u8g2->height;
   xx -= y;
   xx--;
+  
   dir ++;
+  dir &= 3;
+  u8g2_draw_hv_line_4dir(u8g2, xx, yy, len, dir);
+}
+
+void u8g2_draw_l90_r2(u8g2_t *u8g2, u8g2_uint_t x, u8g2_uint_t y, u8g2_uint_t len, uint8_t dir)
+{
+  u8g2_uint_t xx, yy;
+
+  yy = u8g2->height;
+  yy -= y;
+  yy--;
+  
+  xx = u8g2->width;
+  xx -= x;
+  xx--;
+  
+  dir +=2;
   dir &= 3;
   u8g2_draw_hv_line_4dir(u8g2, xx, yy, len, dir);
 }
@@ -105,6 +139,7 @@ void u8g2_draw_l90_r1(u8g2_t *u8g2, u8g2_uint_t x, u8g2_uint_t y, u8g2_uint_t le
 /*============================================*/
 const u8g2_cb_t u8g2_cb_r0 = { u8g2_update_dimension_r0, u8g2_draw_l90_r0 };
 const u8g2_cb_t u8g2_cb_r1 = { u8g2_update_dimension_r1, u8g2_draw_l90_r1 };
+const u8g2_cb_t u8g2_cb_r2 = { u8g2_update_dimension_r2, u8g2_draw_l90_r2 };
